Service: Remove run function from Listen trait
This commit modifies the Listen trait for it only to return a trait
object that is Read and Write. It matches better with the design where
the front end handler is only given a Read and a Write.
The main, infinite loop is now placed in the main function and the
listener no longer holds ownership of the front end handler.
Modifies the CI script to make the tests pass.

Changes provided by:
* Ionut Mihalcea (ionut.mihalcea@arm.com)
* Hugues de Valon (hugues.devalon@arm.com)

Change-Id: I82bbb1eca8042be0d243c2e3d8629745e9b37610
Signed-off-by: Hugues de Valon <hugues.devalon@arm.com>