Service: Add persistence of key triple mappings
This commit adds the key triple to key ID persistence in the On-Disk Key
ID Manager. It adds the new, save and delete methods to be able to
load/save and delete key mappings from/to disk.
The ManageKeyID trait has been modified to now return a Result as there
might be errors when interacting with the persistent device.
The KeyTriple type has been modified to own the application name,
provider and key name data.

Changes provided by:
* Ionut Mihalcea (ionut.mihalcea@arm.com)
* Hugues de Valon (hugues.devalon@arm.com)

Change-Id: Icef07ba9872136b18255cdf1ee5c1b5c7905b227
Signed-off-by: Hugues de Valon <hugues.devalon@arm.com>