Add SIGHUP signal handling to reload configuration
Following daemon(7), it is recommended that systemd daemons reload the
configuration files when SIGHUP is received. This commit implements that
by dropping the Parsec structures and re-instanciating them with the new
configuration.
Adds a Drop trait implementation on the providers for them to
unitialize/free the memory structures of their backends when they are
dropped.
Modifies the test script and CI workflow to use reloading.

Signed-off-by: Hugues de Valon <hugues.devalon@arm.com>