Modify Travis script and removes cross compilation
Modifies the Dockerfile and removes the all-providers tests so that the
tests pass on Aarch64.
When compiling natively, Cargo still uses the linker in .cargo/config
which is a cross-compilation linker and causes problems. Removes the
cross-compilation feature using cross altogether as it is not a strong
use-case for now.
Unites all test scripts into one.

Signed-off-by: Hugues de Valon <hugues.devalon@arm.com>