Improve digest handling in PKCS11 provider
This commit aims to improve our handling of digest data within the
PKCS11 provider. Previously we had harcoded the prefix to the actual
digest, assuming the digest was SHA256 - we are now generating a proper
DigestInfo structure that can be serialized to a ASN.1 DER encoding.

The reason for handling this data ourselves in the PKCS11 provider is
because the PKCS11 interface expects users to provide the formatting.
All other providers only take the digest data and generate the metadata
themselves.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>