Add check to prevent the Parsec service from running as root
This commit introduces a guard at the top of `main` to ensure that
Parsec is not being run as root. Disallowing the Parsec service from
running as root reduces attack surface in line with the principle of
least privilege.

This behaviour can be overridden by setting the `PARSEC_ALLOW_ROOT`
environment variable.

Signed-off-by: Joe Ellis <joe.ellis@arm.com>