Update attribute handling in PKCS11 provider
This commit brings several improvements to the PKCS11 provider.

It updates the way we handle parameters for the PKCS11
provider, making the process of adding support for new key types more
straight forward.
It creates abstractions over PKCS11 types to safely handle some of the
functionality that the pkcs11 crate requires.
It adds asymmetric encryption and decryption support for the PKCS11
provider.
It improves the way we handle test distribution, so each test is run
only for the providers that support that functionality. Since SoftHSM
does not support RSA OAEP with anything but SHA1, a new test was added
for that.

Co-authored-by: Ionut Mihalcea <ionut.mihalcea@arm.com>
Co-authored-by: Hugues de Valon <hugues.devalon@arm.com>

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>