Add support for testing Unix peer credentials authenticator
Certain features require a nightly build of the Rust toolchain. For
example, the peer credentials authenticator currently (Sep 2020)
requires the nightly Rust compiler, because the `peer_cred`
functionality it depends on is only available in the nightly builds.
Nonetheless, we still want to be able to test this functionality. This
commit adds a section to the CI script to allow us to test these sorts
of features with the nightly Rust toolchain.

Signed-off-by: Joe Ellis <joe.ellis@arm.com>