Cherry-pick Rust standard library peer credentials feature into Parsec
The code introduced in this patch has been cherry-picked from the
following PR:

    https://github.com/rust-lang/rust/pull/75148

At the time of writing (16/09/20), this patch is in the nightly Rust
channel. To avoid needing to use the nightly compiler to build Parsec,
this patch includes the change from the standard library to allow us to
use this feature 'early'.

Once the feature hits the stable branch, it should be safe to revert
this commit with `git revert`.

Signed-off-by: Joe Ellis <joe.ellis@arm.com>