Create KeyInfoManagerClient
This commit creates the KeyInfoManagerClient that bridges between the
shared KIM and the providers using it. The new structure is also
deployed across all providers, replacing their internal mechanisms for
handling key information. The mechanisms for generating key IDs remain
in place as they are provider dependent and not generated by the
manager.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>