Enable coverage testing for TS provider
This commit enables collection of code coverage results for the TS
provider, since current changes in the way testing is done means there
is no service restarting, which was triggering the problems before.
The nightly run manual toggle also has an added field, allowing the
tests to be run against specific revision numbers.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>