Execute e2e tests with an old version of client
This is to ensure that the Parsec service does not break for old
clients.
This commit also specifies an exact version of libclang and Mbed Crypto.
Adds build tests for specific versions of the Rust compiler.
Adds a test checking that it's possible to generate a key wiht no flags
set.

Signed-off-by: Hugues de Valon <hugues.devalon@arm.com>