Make cross-compilation run on release version
This commit changes the command for cross compilation building to work
on the release version of the binary instead. This is because a recent
change in the way process liveness is done refers to the release path in
particular.

The version of Rust used to run the coverage tests is also changed to
statically track 1.51.0 to overcome some compilation errors. The SPIFFE
authenticator is also excluded from the tests, as it was proving
problematic during compilation.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>