Adjust TS provider linking
This commit changes the libraries we link to for the TS provider to
work. Namely, it removes many of the previously linked libraries that we
used for testing. It seems even for test builds, there is no reason to
link the other libraries - simply linking to `libts.so` is enough.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>