Fix code coverage reports
This commit fixes our code coverage reports for Parsec. The issue was
found to be the use of the `--release` flag for running the
tarpaulin-instrumented service binary. This flag seems to make
`tarpaulin` lose track of code-to-instruction mapping and thus lose a
lot of code coverage data, maybe due to aliasing.

The `--release` flag was removed and the `ci.sh` script modified to
ensure builds continue to work.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>