Update provider to use new version fo TransKeyCtx
This commit updates the TPM provider to use the updated version of the
TransientKeyContext which handles key material in the form of public and
private fields (instead of an encrypted context). A migration mechanism
is also added, to aid in the recovery of keys stored as contexts.
It also adds a replace_key_info method to the KIM client, which allows
the above-mentioned migration to be done in a data-race safe way.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>