Expand support for importing public keys for TPM
This commit expands the functionality of the TPM provider by allowing it
to import public ECC keys and RSA keys intended for asym encryption.

Provider restrictions have also been lifted on some of the importing and
signing/verification tests, allowing them to be run against the TPM
and/or Trusted Services provider.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>