Bump sd-notify to 0.4.1
In order to package Parsec in Debian we need to package up the
dependencies. One dependency is sd-notify which, when updated, has a
breaking API change.

Specifically, listen_fds() now returns an iterator of fds rather than a
count of fds. Also SD_LISTEN_FDS_START is no longer exposed publicly and
is instead the first value in the iterator returned by listen_fds().

Signed-off-by: Steve Capper <steve.capper@arm.com>