test:docker: Create a common group for parsec users
Motive: The latest rust release (1.65) seems to handle the registry
differently.
It needs to write permissions to the registry for each command.
Even if the package is in the registry.

Solution: Create a common group for parsec clients users which grant them
r/w access to the registry (files in general).
Set the group owner of /opt/rust/registry to the common group and
set permissions to 775. i.e. "drwxrwxr-x 1 root parsec-clients registry"

Signed-off-by: Mohamed Omar Asaker <mohamed.omarasaker@arm.com>