Use ubuntu 22.04 in docker image
The cross compilation of the latest psa-crypto sys crate fails for
aarch64 target on a ubuntu 18.04 machine. This is due to the 0.63.0
version of bindgen in psa-crypto-sys. The latest bindgen crate has
a check which compares size_t, ssize_t with target pointer and aborts
if they are unequal when size_t_is_usize flag is set. In ubuntu 18.04
these are read as 4 and 8 bytes respectively which is wrong. This is
an issue with the c library mismatchs of target toolchains in 18.04.
Hence we are updating the docker test image to 22.04. Also, 18.04 is
reaching end of life in April 2023.

This commit involves the following changes:
- Use ubuntu 22.04 as base image
- Dependency packages missing in 22.04 have been added
- Tpm library versions have been bumped as 22.04 has openssl > 3.0
- Older version libts used to generate keys has been removed as
  the missing dependencies for it causes compilation issues.
  Instead we use parsec 1.0.0 which uses a libts version which is
  compatible with 22.04 for generating keys for both SQLite and
  OnDisk KIM.

Signed-off-by: Gowtham Suresh Kumar <gowtham.sureshkumar@arm.com>

 # OnDisk KIM.