e2e_tests/stress.rs: Add a workaround for spurious test failures
stress_tests fail sporadically with the error:

 'An invalid signature or a tampering detection should be the only
  reasons of the verification failing. Status returned:
  PsaErrorInvalidArgument.'

The verification signature should not be failing with
PsaErrorInvalidArgument for invalid signatures. So there may be a
problem with either threading or a bug in Mbed Crypto.
This bug is causing several sporadic test failures in our CI, as the
invalid signatures are random.

 * Accept the PsaErrorInvalidArgument error as a workaround until
   the issue is resolved.

Please see opened issue

Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>