.cargo/config.toml: remove
When compiling on an aarch64 host, the override of the linker to the
cross "compiler" (uses the gcc frontend) causes a build failure. Theirs
no reason to set these globally for all developers as their build host
may be different as well as their toolchain preference and naming
conventions.

Instead override the target specific linker variable in a --config
option string in the cross compilation tests.

Signed-off-by: Bill Roberts <bill.roberts@arm.com>