tpm: Provider build: Check Stored Public Information of Root Key
The Public Part of the Root Key in the TPM generated by the tss-esapi
layer in the TPM can be checked with a previously public part stored
internally.

Up until this commit, the storing of that information is not done
in the code: This commit only covers the part of checking the Public
Part, not of storing it.

Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>