on_disk_manager: Store Internally generated keys
Keys should be marked as internal or not in the OnDiskManager when
stored.

The object that stores the Key Information (PasswordContext),
however, does not have the capacity to store the Auth information.
When a similar problem arose, a LegacyPasswordContext was created.

This commit avoids having to create a new PasswordContext and allows
not to treat internal keys like any other keys so as not to expose
them.

Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>